backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0,<3.0.0
python-dotenv>=1.0.0
requests>=2.30.0
praw>=7.0.0